Refactored components to use functions instead of classes and started adding styled-components instead of using css style sheets
