+++ src/components/About/styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+import { Section } from '@bootstrap-styled/v4';
+
+
+export const SectionAbout = styled(Section)`
+    background-color: #DABFE3;
+`
+
+export const Lead = styled.p`
+    font-family: 'Comfortaa', cursive;
+`
+export const About = styled.h1`
+    color:#FF3399;
+
+`
