--- src/components/About/styles.css
@@ -1,12 +1,0 @@
-
-#about {
-    background-color: #DABFE3;
-}
-
-.text-pink
-{
-    color:#FF3399;
-}
-.lead{
-    font-family: 'Comfortaa', cursive;
-}
