+++ src/pages/Home/index.js
@@ -1,19 +1,23 @@
-import React, { Component } from "react";
+import React from "react";
 import BootstrapProvider from '@bootstrap-styled/provider/lib/BootstrapProvider';
+
 // import sections
 import Topbar from '../../components/Topbar';
 import Hero from '../../components/Hero';
 import About from '../../components/About';
-class Home extends Component {
-    render() {
-      return (
-        <BootstrapProvider>
-          <Topbar></Topbar>
-          <Hero></Hero>
-          <About></About>
-        </BootstrapProvider>
-      );
-    }
-  }
+
+function Home() {
+  return (
+
+    <BootstrapProvider>
+
+      <Topbar/>
+      <Hero/>
+      <About/>
+
+    </BootstrapProvider>
+  );
+}
+
    
 export default Home;
